software/requirements.txt
numpy
matplotlib
scipy
pyserial